+++ gulpfile.js
@@ -2,11 +2,13 @@ var gulp = require("gulp"),
 livereload = require('gulp-livereload'),
 concat = require("gulp-concat");
 
+var deploymentFolder = './deployment/';
+
 gulp.task('default', ['watch']);
 
 gulp.task('watch', function() {
 	livereload.listen();
-	gulp.watch('**/*.css', ['reload']);
+	gulp.watch('**/*.css', ['reload', 'copy-css']);
 	gulp.watch('**/*.js', ['concat', 'reload']);
 	gulp.watch('**/*.html', ['reload']);
 });
@@ -27,5 +29,11 @@ sketchFiles.forEach(function(fileName, index) {
 gulp.task('concat', function() {
 	gulp.src(sketchFiles)
 	.pipe(concat('concat.js'))
-	.pipe(gulp.dest('./concat/'));
+	.pipe(gulp.dest(deploymentFolder));
+});
+
+
+gulp.task('copy-css', function() {
+	gulp.src('style.css')
+	.pipe(gulp.dest(deploymentFolder));
 });