+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+### Added
+- Copying style.css into deployment folder
+
+### Changed
+- Better boilerplate.
 
 ## [0.2.0] - 2017-04-29
 ### Added